Move logging-related files to config package.
